requirements.txt
# Core ML/DL Libraries
tensorflow>=2.8.0
torch>=1.12.0
scikit-learn>=1.1.0
numpy>=1.21.0
pandas>=1.4.0

# Deep Learning & Neural Networks
keras>=2.8.0
sonnet>=2.0.0

# Data Processing & Analysis
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Financial Libraries
yfinance>=0.1.70
alpha-vantage>=2.3.1
ta>=0.10.1

# Web Framework
flask>=2.2.0
flask-cors>=3.0.10

# Utilities
requests>=2.28.0
python-dotenv>=0.19.0
tqdm>=4.64.0
PyYAML>=6.0

# Optional: GPU Support
# tensorflow-gpu>=2.8.0
# torch-cuda>=1.12.0 